lib: merge lur-cache and redis-cache module
