+++ lib/index.js
@@ -6,8 +6,7 @@ const logger = require('./utils/logger');
 const onerror = require('./middleware/onerror');
 const header = require('./middleware/header');
 const utf8 = require('./middleware/utf8');
-const memoryCache = require('./middleware/lru-cache');
-const redisCache = require('./middleware/redis-cache');
+const cache = require('./middleware/cache');
 const parameter = require('./middleware/parameter');
 const template = require('./middleware/template');
 const favicon = require('koa-favicon');
@@ -65,57 +64,9 @@ app.use(template);
 app.use(parameter);
 
 // 2 cache
-if (config.cacheType === 'memory') {
-    app.use(
-        memoryCache({
-            app: app,
-            expire: config.cacheExpire,
-            ignoreQuery: true,
-        })
-    );
-} else if (config.cacheType === 'redis') {
-    app.use(
-        redisCache({
-            app: app,
-            expire: config.cacheExpire,
-            ignoreQuery: true,
-            redis: config.redis,
-            onerror: (e) => {
-                logger.error('Redis error: ', e);
-            },
-            onconnect: () => {
-                logger.info('Redis connected.');
-            },
-        })
-    );
-} else {
-    app.context.cache = {
-        get: () => null,
-        set: () => null,
-    };
-}
-app.context.cache.tryGet = async function(key, getValueFunc, maxAge = 24 * 60 * 60) {
-    let v = await this.get(key);
-    if (!v) {
-        v = await getValueFunc();
-        this.set(key, v, maxAge);
-    } else {
-        let parsed;
-        try {
-            parsed = JSON.parse(v);
-        } catch (e) {
-            parsed = null;
-        }
-        if (parsed) {
-            v = parsed;
-        }
-    }
-
-    return v;
-};
+app.use(cache(app));
 
 // router
-
 app.use(mount('/', router.routes())).use(router.allowedMethods());
 
 // routes the require authentication
