+++ lib/routes/zhihu/daily.js
@@ -24,7 +24,7 @@ module.exports = async (ctx) => {
                 link: 'https://news-at.zhihu.com/story/' + story.id,
             };
             const key = 'daily' + story.id;
-            const value = await ctx.cache.get(key, true);
+            const value = await ctx.cache.get(key);
 
             if (value) {
                 item.description = value;
@@ -37,7 +37,7 @@ module.exports = async (ctx) => {
                     },
                 });
                 item.description = utils.ProcessImage(storyDetail.data.body.replace(/<div class="meta">([\s\S]*?)<\/div>/g, '<strong>$1</strong>').replace(/<h2.*?>/g, ''));
-                ctx.cache.set(key, item.description, 24 * 60 * 60, true);
+                ctx.cache.set(key, item.description, 24 * 60 * 60);
             }
 
             return Promise.resolve(item);
